Create a duckdb cdm_reference from a dataset.

Note these are temporal cdm objects that won't be accessible after disconnecting, if you want to create a duckdb database use exportDatasetToDuckdb().

Usage

cdmFromDataset(
  datasetName = "GiBleed",
  cdmSchema = "main",
  cdmPrefix = "",
  writeSchema = "main",
  writePrefix = ""
)

Arguments

datasetName

Name of the dataset. See availableDatasets() for possibilities.

cdmSchema

A string that indicates the cdm schema.

cdmPrefix

A string that prefixes the tables in the cdm schema.

writeSchema

A string that indicates the write schema.

writePrefix

A string that prefixes the tables in the write schema.

Value

A cdm_reference object.

Examples

library(OmopDatasets)

datasetsFolder(tempdir())
#> ℹ If you want to create a permanent `OMOP_DATASETS_FOLDER` write the following
#>   in your `.Renviron` file:
#> 
#>   OMOP_DATASETS_FOLDER="/tmp/RtmpZCzHJ6"
downloadDataset(datasetName = "GiBleed")
cdm <- cdmFromDataset(datasetName = "GiBleed")
cdm
#> 
#> ── # OMOP CDM reference (duckdb) of GiBleed ────────────────────────────────────
#> • omop tables: person, observation_period, visit_occurrence, visit_detail,
#> condition_occurrence, drug_exposure, procedure_occurrence, device_exposure,
#> measurement, observation, death, note, note_nlp, specimen, fact_relationship,
#> location, care_site, provider, payer_plan_period, cost, drug_era, dose_era,
#> condition_era, metadata, cdm_source, concept, vocabulary, domain,
#> concept_class, concept_relationship, relationship, concept_synonym,
#> concept_ancestor, source_to_concept_map, drug_strength
#> • cohort tables: -
#> • achilles tables: -
#> • other tables: -